Add gRPC health checks to registration service
It turns out that simply adding io.grpc:grpc-services as a depdendency
is enough for micronaut to automatically set up gRPC health checks.

By default the check succeeds when the service is running and fails
otherwise. We can extend this using the HealthStatusManager as described
in this guide:

https://micronaut-projects.github.io/micronaut-grpc/latest/guide/

I tested this in dev mode using grpcurl and was able to get a valid
response when the service was running.