Only consider first attempts for stats
Since we use a different selection strategy when making multiple
attempts on the same session, bias is introduced in the data. If a
provider mostly recieves only "second" attempts, (which are likely
for numbers that are facing deliverability problems to begin with)
their performance might be worse than if they were recieving a random
sample of requests.

For the purposes of adaptive routing choices, only use data from the
first attempt in a session.